requirements.txt
absl-py==2.3.1
agent-starter-pack==0.22.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
attrs==25.3.0
authlib==1.6.5
backoff==2.2.1
binaryornot==0.4.4
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cookiecutter==2.6.0
cryptography==46.0.2
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.118.0
fastuuid==0.13.5
filelock==3.19.1
-e file:///Users/allancastellanos/agentemerc
frozenlist==1.7.0
fsspec==2025.9.0
google-adk==1.19.0
google-api-core==2.28.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.126.1
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==3.5.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.52.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.76.3
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.16.0
mdurl==0.1.2
multidict==6.6.4
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.3
openai==2.1.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.9.0a0
opentelemetry-exporter-gcp-monitoring==1.9.0a0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-google-genai==0.4b0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-genai==0.2b0
packaging==25.0
pandas==2.3.3
pgvector==0.4.2
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.11.0
pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rouge-score==0.1.2
rpds-py==0.27.1
rsa==4.9.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
